Pruebas temperatura + identificar sensores

+++ Ecoembes/src/main/java/es/deusto/sd/auctions/facade/EcoembesController.java
@@ -162,7 +162,7 @@ public class EcoembesController {
     @GetMapping("/plantas/{id_planta}/{fecha}")
     public ResponseEntity<Double> get_capacidad_planta_fecha(
             @Parameter(name = "id_planta", description = "id de la planta", required = true, example = "00001")
-            @PathVariable("id_planta") long planta,
+            @PathVariable("id_planta") int planta,
             @Parameter(name = "fecha", description = "fecha de la que quiero la capacidad ", required = true, example = "01-01-2025")
             @PathVariable("fecha") String fecha,
             @Parameter(description = "Token de autenticación del usuario", required = true, example = "abc123xyz")
